1) fix texture caching issue (ie. MP1 "Main menu" fps drop). 2) "fix" Pokemon Coloseum font issue in SafeTextureCache mode. We should kill the unsafe one. TODO: get why it works :p 3) readjust FAKE_GP_WATCHDOG_PERIOD for omega's ZTP to boot in DC. Too late to split, sorry.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@1871 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/VideoCommon/Src/TextureDecoder.h
@@ -78,7 +78,7 @@ enum PC_TexFormat
 PC_TexFormat TexDecoder_Decode(u8 *dst, const u8 *src, int width, int height, int texformat, int tlutaddr, int tlutfmt);
 
 u32 TexDecoder_GetSafeTextureHash(const u8 *src, int width, int height, int texformat, u32 seed=0);
-u32 TexDecoder_GetTlutHash(const u16 *src, int len);
+u32 TexDecoder_GetTlutHash(const u8* src, int len);
 
 void TexDecoder_SetTexFmtOverlayOptions(bool enable, bool center);
